# app/doc/openapi.yaml
openapi: 3.0.0
info:
  title: 'API documentation for blog example.'
  description: 'Implementation of Swagger with in Laravel'
  contact:
    email: sergmoro1@ya.ru
  license:
    name: MIT
    url: 'https://mit-license.org/'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8080'
    description: 'Blog API Server'
paths:
  /api/posts:
    get:
      tags:
        - Posts
      summary: 'Get list of posts'
      description: 'Returns list of posts'
      operationId: indexPost
      parameters:
        -
          name: 'tags[]'
          in: query
          description: 'tag to filter by'
          required: false
          style: form
          schema:
            type: array
            items:
              type: string
        -
          name: limit
          in: query
          description: 'maximum number of results to return'
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    put:
      tags:
        - Posts
      summary: 'Update post'
      description: 'Update existing post'
      operationId: updatePost
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostOnly'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: 'true' }
                  message: { type: string, example: 'Post updated.' }
                type: object
    post:
      tags:
        - Posts
      summary: 'Create post'
      description: 'Create new post'
      operationId: createPost
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostOnly'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: 'true' }
                  message: { type: string, example: 'Post created.' }
                type: object
  '/api/posts/{id}':
    get:
      tags:
        - Posts
      summary: 'Get post'
      description: 'Get post by ID'
      operationId: getPost
      parameters:
        -
          name: id
          in: query
          description: 'Post ID'
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
    delete:
      tags:
        - Posts
      summary: 'Delete post'
      description: 'Delete post by ID'
      operationId: deletePost
      parameters:
        -
          name: id
          in: query
          description: 'Post ID'
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: 'true' }
                  message: { type: string, example: 'Post deleted.' }
                type: object
  /api/tags:
    get:
      tags:
        - Tags
      summary: 'Get list of tags'
      description: 'Returns list of tags'
      operationId: indexTag
      parameters:
        -
          name: search
          in: query
          description: 'search string'
          required: false
          schema:
            type: string
        -
          name: limit
          in: query
          description: 'maximum number of results to return'
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
    put:
      tags:
        - Tags
      summary: 'Update tag'
      description: 'Update existing tag'
      operationId: updateTag
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagOnly'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: 'true' }
                  message: { type: string, example: 'Tag updated.' }
                type: object
    post:
      tags:
        - Tags
      summary: 'Create tag'
      description: 'Create new tag'
      operationId: createTag
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagOnly'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: 'true' }
                  message: { type: string, example: 'Tag created.' }
                type: object
  '/api/tags/{id}':
    get:
      tags:
        - Tags
      summary: 'Get tag'
      description: 'Get tag by ID'
      operationId: getTag
      parameters:
        -
          name: id
          in: query
          description: 'Tag ID'
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
    delete:
      tags:
        - Tags
      summary: 'Delete tag'
      description: 'Delete tag by ID'
      operationId: deleteTag
      parameters:
        -
          name: id
          in: query
          description: 'Tag ID'
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: 'true' }
                  message: { type: string, example: 'Tag deleted.' }
                type: object
components:
  schemas:
    PostOnly:
      properties:
        id:
          type: integer
          example: 1
        status:
          type: string
          enum:
            - draft
            - published
            - archived
          example: draft
        title:
          type: string
          example: 'Genius title'
        excerpt:
          type: string
          example: 'Excerpt of Content'
        Content:
          type: string
          example: 'Long Content'
        created_at:
          type: string
          format: date-time
          example: '2022-07-04T16:31:00+0300'
        updated_at:
          type: string
          format: date-time
          example: '2022-07-04T16:31:00+0300'
      type: object
    Post:
      properties:
        id:
          type: integer
          example: 1
        status:
          type: string
          enum:
            - draft
            - published
            - archived
          example: draft
        title:
          type: string
          example: 'Genius title'
        excerpt:
          type: string
          example: 'Excerpt of Content'
        Content:
          type: string
          example: 'Long Content'
        Tags:
          type: array
          items:
            $ref: '#/components/schemas/TagOnly'
        created_at:
          type: string
          format: date-time
          example: '2022-07-04T16:31:00+0300'
        updated_at:
          type: string
          format: date-time
          example: '2022-07-04T16:31:00+0300'
      type: object
    TagOnly:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: php
        created_at:
          type: string
          format: date-time
          example: '2022-07-04T16:31:00+0300'
        updated_at:
          type: string
          format: date-time
          example: '2022-07-04T16:31:00+0300'
      type: object
    Tag:
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: php
        Posts:
          type: array
          items:
            $ref: '#/components/schemas/PostOnly'
        created_at:
          type: string
          format: date-time
          example: '2022-07-04T16:31:00+0300'
        updated_at:
          type: string
          format: date-time
          example: '2022-07-04T16:31:00+0300'
      type: object
  securitySchemes:
    basicAuth:
      type: http
      description: 'Basic authentication by email & password.'
      name: Authorization
      in: header
      scheme: basic